requirements.txt
absl-py==1.2.0
astunparse==1.6.3
attrs==22.1.0
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
dataclasses==0.6
flatbuffers==23.5.26
gast==0.4.0
google-auth-oauthlib==0.4.6
google-auth==2.19.1
google-pasta==0.2.0
grpcio==1.54.2
h5py==3.8.0
idna==3.4
importlib-metadata==6.6.0
keras==2.11.0
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.3
mediapipe==0.8.3.1
numpy==1.23.3
oauthlib==3.2.2
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==23.1
pdf2image==1.16.3
Pillow==9.3.0
pip==21.3.1
protobuf==3.15.6
pyasn1-modules==0.3.0
pyasn1==0.5.0
PyMuPDF==1.18.17
pypdf2==2.12.1
python-dotenv==1.0.0
reportlab==3.6.12
requests-oauthlib==1.3.1
requests==2.31.0
rsa==4.9
setuptools==60.2.0
six==1.16.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboard==2.11.2
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow==2.11.0
termcolor==2.3.0
tkPDFViewer==0.1
typing-extensions==4.5.0
urllib3==1.26.16
Werkzeug==2.3.5
wheel==0.37.1
wrapt==1.15.0
zipp==3.15.0
